remove name check in Leaflet constructor
If you are working on a Vue project and mappa is a dependency the
vue-cli bin command will uglify/minify your code. As a part of this
process all of the classes in exisiting Javascript files get renamed.
When this happens the check to see if this.constructor.name is equal to
'Leaflet' will fail as the class may have been renamed.

After examining the Leaflet class as well as well as other tile
providers I believe that this check itself is unnecessary and by
removing it we also solve the above issue.